+++ db/migrate/20171110003855_create_income.rb
@@ -1,10 +1,10 @@
 class CreateIncome < ActiveRecord::Migration[5.1]
   def change
     create_table :incomes do |t|
-      t.integer :earnedIncome
-      t.integer :netInvestedAssets
-      t.integer :taxRate
-      t.integer :roi
+      t.float :earnedIncome
+      t.float :netInvestedAssets
+      t.float :taxRate, precision: 2
+      t.float :roi, precision: 2
     end
   end
 end
